refactor(marketing): modularize live site announcements.
Split the monolith into lib helpers, focused UI pieces, and small hooks;
move shared types out of the client component and harden fetch/consent handling.
