+++ lib/marketing/siteAnnouncements/constants.ts
@@ -1,0 +1,15 @@
+export const ANNOUNCE_BANNER_HEIGHT_VAR = "--announce-banner-h";
+export const ANNOUNCE_BANNER_HEIGHT_PX = "36px";
+
+export const DISMISS_STORAGE_PREFIX = "fixera-announce-dismiss:";
+export const PREVIEW_DURATION_MS = 5000;
+
+export const GEO_COUNTRY_COOKIE = "fixera_geo_country";
+export const GEO_LOCALE_COOKIE = "fixera_geo_locale";
+
+export const SKIP_PATH_PREFIXES = [
+  "/admin",
+  "/login",
+  "/signup",
+  "/register",
+] as const;
