+++ lib/marketing/siteAnnouncements/cookies.ts
@@ -1,0 +1,7 @@
+/** Read a document cookie by name. Returns null when missing or not in browser. */
+export function readCookie(name: string): string | null {
+  if (typeof document === "undefined") return null;
+  const escaped = name.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+  const match = document.cookie.match(new RegExp(`(?:^|; )${escaped}=([^;]*)`));
+  return match ? decodeURIComponent(match[1]) : null;
+}
